wip: Login into Esal
